+++ streamrip/client/__init__.py
@@ -1,9 +1,9 @@
 from .client import Client
-from .deezer_client import DeezerClient
+from .deezer import DeezerClient
 from .downloadable import BasicDownloadable, Downloadable
-from .qobuz_client import QobuzClient
-from .soundcloud_client import SoundcloudClient
-from .tidal_client import TidalClient
+from .qobuz import QobuzClient
+from .soundcloud import SoundcloudClient
+from .tidal import TidalClient
 
 __all__ = [
     "Client",
