+++ streamrip/client/tidal.py
