+++ streamrip/media/__init__.py
@@ -1,21 +1,25 @@
 from .album import Album, PendingAlbum
 from .artist import Artist, PendingArtist
+from .artwork import remove_artwork_tempdirs
 from .label import Label, PendingLabel
-from .media import Media
+from .media import Media, Pending
 from .playlist import PendingPlaylist, PendingPlaylistTrack, Playlist
-from .track import PendingTrack, Track
+from .track import PendingSingle, PendingTrack, Track
 
 __all__ = [
-    "Album",
-    "Artist",
-    "Label",
     "Media",
+    "Pending",
+    "Album",
     "PendingAlbum",
+    "Artist",
     "PendingArtist",
+    "Label",
     "PendingLabel",
-    "PendingPlaylist",
-    "PendingPlaylistTrack",
-    "PendingTrack",
     "Playlist",
+    "PendingPlaylist",
     "Track",
+    "PendingTrack",
+    "PendingPlaylistTrack",
+    "PendingSingle",
+    "remove_artwork_tempdirs",
 ]
