+++ streamrip/rip/prompter.py
@@ -5,13 +5,9 @@ from getpass import getpass
 
 from click import launch, secho, style
 
-from .client import Client
-from .config import Config
-from .deezer_client import DeezerClient
-from .exceptions import AuthenticationError, MissingCredentials
-from .qobuz_client import QobuzClient
-from .soundcloud_client import SoundcloudClient
-from .tidal_client import TidalClient
+from ..client import Client, DeezerClient, QobuzClient, SoundcloudClient, TidalClient
+from ..config import Config
+from ..exceptions import AuthenticationError, MissingCredentials
 
 
 class CredentialPrompter(ABC):
